requirements.txt
torch
torchvision
numpy
scikit-learn
matplotlib
transformers
pandas
pillow